Extract execution-layer system-invariant checks into an InvariantChecker module
