+++ sui-execution/latest/sui-adapter/src/execution_engine.rs
@@ -37,7 +37,6 @@ mod checked {
 
     use crate::static_programmable_transactions as SPT;
     use crate::sui_types::gas::SuiGasStatusAPI;
-    use crate::type_layout_resolver::TypeLayoutResolver;
     use crate::{gas_charger::GasCharger, temporary_store::TemporaryStore};
     use move_core_types::ident_str;
     use move_core_types::language_storage::TypeTag;
@@ -214,54 +213,6 @@ mod checked {
         }
     }
 
-    /// Compute the per-`(address, type)` funds-accumulator reservation budget consumed by
-    /// `TemporaryStore::check_address_balance_changes`. Today every funds accumulator is a
-    /// `Balance<T>`, but the `(address, TypeTag)` keying lets this generalize as more
-    /// accumulator types are added. Sources:
-    /// - PTB `FundsWithdrawalArg`s for any supported accumulator type (sender or sponsor as
-    ///   owner).
-    /// - Gas paid entirely from address balance (credits `(gas_owner, Balance<SUI>)`).
-    /// - Gas-data entries with coin-reservation digests (also credit `(gas_owner, Balance<SUI>)`).
-    fn compute_input_reservations(
-        transaction_kind: &TransactionKind,
-        gas_data: &GasData,
-        transaction_signer: SuiAddress,
-    ) -> BTreeMap<(SuiAddress, TypeTag), u64> {
-        use sui_types::balance::Balance;
-        use sui_types::gas_coin::GAS;
-        use sui_types::transaction::{Reservation, WithdrawFrom, is_gas_paid_from_address_balance};
-
-        let mut reservations: BTreeMap<(SuiAddress, TypeTag), u64> = BTreeMap::new();
-        let sui_balance_type = Balance::type_tag(GAS::type_tag());
-
-        for arg in transaction_kind.get_funds_withdrawals() {
-            let owner = match arg.withdraw_from {
-                WithdrawFrom::Sender => transaction_signer,
-                WithdrawFrom::Sponsor => gas_data.owner,
-            };
-            let Reservation::MaxAmountU64(reservation) = arg.reservation;
-            *reservations
-                .entry((owner, arg.type_arg.to_type_tag()))
-                .or_insert(0) += reservation;
-        }
-
-        if is_gas_paid_from_address_balance(gas_data, transaction_kind) {
-            *reservations
-                .entry((gas_data.owner, sui_balance_type.clone()))
-                .or_insert(0) += gas_data.budget;
-        }
-
-        for entry in &gas_data.payment {
-            if let Ok(parsed) = ParsedDigest::try_from(entry.2) {
-                *reservations
-                    .entry((gas_data.owner, sui_balance_type.clone()))
-                    .or_insert(0) += parsed.reservation_amount();
-            }
-        }
-
-        reservations
-    }
-
     #[allow(clippy::type_complexity)]
     #[instrument(name = "tx_execute_to_effects", level = "debug", skip_all)]
     pub fn execute_transaction_to_effects<Mode: ExecutionMode>(
@@ -396,8 +347,10 @@ mod checked {
             && is_gasless_transaction(&gas_data, &transaction_kind);
         let is_epoch_change = transaction_kind.is_end_of_epoch_tx();
 
-        let input_reservations =
-            compute_input_reservations(&transaction_kind, &gas_data, transaction_signer);
+        // Cache the transaction-derived invariant-check inputs (reservation budget, advance-epoch
+        // mint/burn, genesis flag) on the store, after the gas-smash filtering above. The
+        // conservation checks and the ownership-invariant check read them from there.
+        temporary_store.set_invariant_inputs(&transaction_kind, &gas_data, transaction_signer);
 
         let (gas_cost_summary, execution_result, timings) = execute_transaction::<Mode>(
             store,
@@ -413,7 +366,6 @@ mod checked {
             execution_params,
             trace_builder_opt,
             is_gasless,
-            &input_reservations,
         );
 
         let status = if let Err(error) = &execution_result {
@@ -444,7 +396,6 @@ mod checked {
                     &sponsor,
                     &gas_charger,
                     &mutable_inputs,
-                    &input_reservations,
                     is_epoch_change,
                 )
                 .unwrap()
@@ -581,7 +532,6 @@ mod checked {
         execution_params: ExecutionOrEarlyError,
         trace_builder_opt: &mut Option<MoveTraceBuilder>,
         is_gasless: bool,
-        input_reservations: &BTreeMap<(SuiAddress, TypeTag), u64>,
     ) -> (
         GasCostSummary,
         Result<Mode::ExecutionResults, Mode::Error>,
@@ -593,8 +543,6 @@ mod checked {
             "No gas charges must be applied yet"
         );
 
-        let is_genesis_tx = matches!(transaction_kind, TransactionKind::Genesis(_));
-        let advance_epoch_gas_summary = transaction_kind.get_advance_epoch_tx_gas_summary();
         let digest = tx_ctx.borrow().digest();
         let withdrawal_reservations =
             if is_gasless && protocol_config.gasless_verify_remaining_balance() {
@@ -708,9 +656,6 @@ mod checked {
             protocol_config,
             enable_expensive_checks,
             &cost_summary,
-            is_genesis_tx,
-            advance_epoch_gas_summary,
-            input_reservations,
         ) {
             // FIXME: we cannot fail the transaction if this is an epoch change transaction.
             result = Err(e);
@@ -719,6 +664,11 @@ mod checked {
         (cost_summary, result, timings)
     }
 
+    /// Run the SUI-conservation and balance-accumulator invariant checks ([`invariants::run_all_checks`])
+    /// against the finalized store. On a violation, recover by dumping all writes, charging gas in
+    /// the aborted state, and re-checking; a surviving double failure means gas charging itself
+    /// mints or burns SUI, which is unrecoverable, so we panic. The checks themselves are read-only;
+    /// the recovery's gas-charging mutations are orchestrated here alongside the main-path charge.
     #[instrument(name = "run_conservation_checks", level = "debug", skip_all)]
     fn run_conservation_checks<Mode: ExecutionMode>(
         temporary_store: &mut TemporaryStore<'_>,
@@ -728,61 +678,36 @@ mod checked {
         protocol_config: &ProtocolConfig,
         enable_expensive_checks: bool,
         cost_summary: &GasCostSummary,
-        is_genesis_tx: bool,
-        advance_epoch_gas_summary: Option<(u64, u64)>,
-        input_reservations: &BTreeMap<(SuiAddress, TypeTag), u64>,
     ) -> Result<(), Mode::Error> {
-        let simple_conservation_checks = protocol_config.simple_conservation_checks();
-        let mut result: Result<(), Mode::Error> = Ok(());
-        if !is_genesis_tx && !Mode::skip_conservation_checks() {
-            let run_checks = |store: &TemporaryStore<'_>| -> Result<(), ExecutionError> {
-                store
-                    .check_sui_conserved(simple_conservation_checks, cost_summary)
-                    .and_then(|()| {
-                        if enable_expensive_checks {
-                            let mut layout_resolver = TypeLayoutResolver::new(
-                                move_vm,
-                                store.protocol_config(),
-                                Box::new(store),
-                            );
-                            store.check_sui_conserved_expensive(
-                                cost_summary,
-                                advance_epoch_gas_summary,
-                                &mut layout_resolver,
-                            )
-                        } else {
-                            Ok(())
-                        }
-                    })
-                    .and_then(|()| {
-                        store.check_address_balance_changes(
-                            store.protocol_config(),
-                            input_reservations,
-                        )
-                    })
-            };
+        let Err(conservation_err) = temporary_store.check_conservation_invariants::<Mode>(
+            move_vm,
+            enable_expensive_checks,
+            cost_summary,
+        ) else {
+            return Ok(());
+        };
 
-            if let Err(conservation_err) = run_checks(temporary_store) {
-                // conservation violated. try to avoid panic by dumping all writes, charging for gas,
-                // re-checking conservation, and surfacing an aborted transaction with an invariant
-                // violation if all of that works.
-                result = Err(conservation_err.into());
-                gas_charger.reset(temporary_store);
-                gas_charger.charge_gas(temporary_store, protocol_config, &mut result);
-                if let Err(recovery_err) = run_checks(temporary_store) {
-                    // if we still fail, it's a problem with gas charging that happens even in the
-                    // "aborted" case — no other option but panic. We will create or destroy SUI
-                    // otherwise (or admit an unauthorized accumulator Split).
-                    panic!(
-                        "SUI conservation fail in tx block {}: {}\nGas status is {}\nTx was ",
-                        tx_digest,
-                        recovery_err,
-                        gas_charger.summary()
-                    )
-                }
-            }
-        } // else, we're in the genesis transaction which mints the SUI supply, and hence does not satisfy SUI conservation, or
-        // we're in the non-production dev inspect mode which allows us to violate conservation
+        // Conservation violated. Try to avoid a panic by dumping all writes, charging for gas in
+        // the aborted state, and re-checking; surface an aborted transaction with the invariant
+        // violation if that works.
+        let mut result: Result<(), Mode::Error> = Err(conservation_err.into());
+        gas_charger.reset(temporary_store);
+        gas_charger.charge_gas(temporary_store, protocol_config, &mut result);
+        if let Err(recovery_err) = temporary_store.check_conservation_invariants::<Mode>(
+            move_vm,
+            enable_expensive_checks,
+            cost_summary,
+        ) {
+            // If we still fail, it's a problem with gas charging that happens even in the
+            // "aborted" case — no other option but panic. We would create or destroy SUI
+            // otherwise (or admit an unauthorized accumulator Split).
+            panic!(
+                "SUI conservation fail in tx block {}: {}\nGas status is {}\nTx was ",
+                tx_digest,
+                recovery_err,
+                gas_charger.summary()
+            )
+        }
         result
     }
